Update README.md
The ATM Interface project is a Java-based application that simulates the functionalities of an Automated Teller Machine (ATM). It provides users with a user-friendly interface to perform basic banking operations such as cash withdrawals, balance inquiries, fund transfers, and account management. This project aims to demonstrate the fundamental concepts of object-oriented programming and software design while providing a practical simulation of real-world banking transactions.